shared threads is done
