+++ src/styles/tailwind.css
@@ -18,6 +18,7 @@
     --light_blue_a700: #048ffd;
     --white_a700: #ffffff;
     --white_a700_19: #ffffff19;
+    --white: #ffffff
 
     /*------Border radius variables------*/
     --radius-xs: 1px;
